pyproject.toml
[project]
name = "inspect-evals-starter-with-policies"
version = "0.1.0"
description = "Minimal eval starter with YAML policies"
requires-python = ">=3.9"
authors = [{name = "Your Team"}]
dependencies = [
    "pyyaml>=6.0",
    "pandas>=2.0",
    "rich>=13.0",
    "regex>=2023.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
